fix fields test examples
