+++ packages/parts/src/routes/4-export/print/mock-data.ts
@@ -242,3 +242,64 @@ export const entries = [
     id: 'cY1McfcmkJyv6EqHeynb',
   },
 ];
+
+export const entriesWithSoundFiles = [
+  {
+    id: '1',
+    lx: 'Without speaker',
+    gl: null,
+  },
+  {
+    id: '2',
+    lx: 'With speaker',
+    gl: null,
+    speaker: 'Anonymus',
+  },
+];
+
+export const entriesWithSemanticDomains = [
+  {
+    id: '1',
+    lx: 'One semantic domain',
+    gl: null,
+    sdn: ['4'],
+  },
+  {
+    id: '2',
+    lx: 'Five semantic domains',
+    gl: null,
+    sdn: ['1', '1.5', '2.1', '2.2', '2.3'],
+  },
+  {
+    id: '3',
+    lx: 'Many semantic domains',
+    gl: null,
+    sdn: [
+      '1.3',
+      '1.7',
+      '2.2',
+      '2.6',
+      '2.8',
+      '3',
+      '4',
+      '5',
+      '5.17',
+      '6',
+      '7',
+      '8',
+      '9',
+      '10',
+      '10.1',
+      '10.3',
+      '10.6',
+      '10.8',
+      '10.9',
+    ],
+  },
+  {
+    id: '4',
+    lx: 'None semantic domains',
+    gl: null,
+    sdn: [],
+  },
+];
