+++ packages/parts/src/routes/4-export/print/printFieldsForTesting.ts
@@ -1,0 +1,61 @@
+import type { IPrintFields } from '@living-dictionaries/types';
+
+export const configureSizesFields: IPrintFields = {
+  alternateOrthographies: true,
+  ph: true,
+  ps: true,
+  gloss: true,
+  example_sentence: true,
+  sr: false,
+  sdn: false,
+  mr: false,
+  in: false,
+  nc: false,
+  pl: false,
+  va: false,
+  di: false,
+  nt: false,
+  image: true,
+  speaker: false,
+  id: false,
+};
+
+export const semanticDomainsField: IPrintFields = {
+  alternateOrthographies: false,
+  ph: false,
+  ps: false,
+  gloss: false,
+  example_sentence: false,
+  sr: false,
+  sdn: true,
+  mr: false,
+  in: false,
+  nc: false,
+  pl: false,
+  va: false,
+  di: false,
+  nt: false,
+  image: false,
+  speaker: false,
+  id: false,
+};
+
+export const speakerField: IPrintFields = {
+  alternateOrthographies: false,
+  ph: false,
+  ps: false,
+  gloss: false,
+  example_sentence: false,
+  sr: false,
+  sdn: false,
+  mr: false,
+  in: false,
+  nc: false,
+  pl: false,
+  va: false,
+  di: false,
+  nt: false,
+  image: false,
+  speaker: true,
+  id: false,
+};
